requirements.txt
streamlit==1.11.0
python-dotenv==0.19.2
azure-ai-formrecognizer==3.2.0
openai==0.27.0
